requirements.txt
streamlit
langchain
langchain-openai
langchain-community
pinecone-client
tiktoken
python-dotenv
docx2txt
pypdf